critical fix in auto-upgrade store from v2 to v3
The old-style write position (first 8 bytes as int64) is NOT relative to beginning of the file, but relative to 8 bytes into the file!
In V3, the write position is relative to beginning of the file, so we have to add 8, otherwise we will overwrite the final 8 bytes of the last store entry.